+++ src/OperationModeNode.hpp
@@ -33,17 +33,20 @@ public:
   void setPoolTemperatureNode(DallasTemperatureNode* node) { _currentPoolTempNode = node; };
   void setSolarTemperatureNode(DallasTemperatureNode* node) { _currentSolarTempNode = node; };
 
-  void  setPoolMaxTemperature(float temp) { _poolMaxTemp = temp; };
+  void  setPoolMaxTemperature(float temp) { _poolMaxTemp = temp; saveState(); };
   float getPoolMaxTemperature() { return _poolMaxTemp; };
 
-  void  setSolarMinTemperature(float temp) { _solarMinTemp = temp; };
+  void  setSolarMinTemperature(float temp) { _solarMinTemp = temp; saveState(); };
   float getSolarMinTemperature() { return _solarMinTemp; };
 
-  void  setTemperatureHysteresis(float temp) { _hysteresis = temp; };
+  void  setTemperatureHysteresis(float temp) { _hysteresis = temp; saveState(); };
   float getTemperatureHysteresis() { return _hysteresis; };
 
-  void         setTimerSetting(TimerSetting setting) { _timerSetting = setting; };
+  void         setTimerSetting(TimerSetting setting) { _timerSetting = setting; saveState(); };
   TimerSetting getTimerSetting() { return _timerSetting; };
+
+  void loadState();
+  void saveState();
 
   enum MODE { AUTO, MANU, BOOST };
   const char* STATUS_AUTO  = "auto";